Split batch
